requirements.txt
aiogram>=3.5,<4
aiocryptopay
python-dotenv
